requirements.txt.txt
python-telegram-bot==13.15
Flask==2.2.5
schedule==1.2.1
matplotlib==3.7.1
pandas==2.0.3
